requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
matplotlib

# api
requests

# utils
ipython
etherscan-python

# ape
eth-ape
ape-foundry
ape-solidity
